+++ src/assets/images/index.js
@@ -1,5 +1,6 @@
 import ImgPattern from './pattern.svg';
 import ImgFoody from './foody.svg';
 import ImgDefault from './imgdefault.png';
+import ImgWelcome from './ilus_welcome.svg';
 
-export {ImgDefault, ImgPattern, ImgFoody};
+export {ImgDefault, ImgPattern, ImgFoody, ImgWelcome};